# ENH add pairwise distances backend for LocalOutlierFactor