--- a/sklearn/metrics/_pairwise_distances_reduction/_argkmin_lrd.pyx.tp
+++ b/sklearn/metrics/_pairwise_distances_reduction/_argkmin_lrd.pyx.tp
@@ -18,13 +18,13 @@ from ._datasets_pair cimport DatasetsPair{{name_suffix}}
 
 cdef class ArgKminLRD{{name_suffix}}(ArgKmin{{name_suffix}}):
     """
-    {{name_suffix}}bit implementation of ArgKminLRD
+    {{name_suffix}}bit implementation of ArgKminLRD. Current
+    implementation supports the fit method of LocalOutlierFactor.
     """
     cdef:
         float64_t[::1] lrd
         float64_t[:, ::1] actual_distances
         intp_t[:, ::1] actual_indices
-        intp_t factor
 
     @classmethod
     def compute(
@@ -36,7 +36,6 @@ cdef class ArgKminLRD{{name_suffix}}(ArgKmin{{name_suffix}}):
         chunk_size=None,
         dict metric_kwargs=None,
         str strategy=None,
-        bint train=False,
         bint return_distance=False,
     ):
         # Use a generic implementation that handles most scipy
@@ -46,7 +45,6 @@ cdef class ArgKminLRD{{name_suffix}}(ArgKmin{{name_suffix}}):
             k=k,
             chunk_size=chunk_size,
             strategy=strategy,
-            train=train,
         )
 
         # Limit the number of threads in second level of nested parallelism for BLAS
@@ -65,7 +63,6 @@ cdef class ArgKminLRD{{name_suffix}}(ArgKmin{{name_suffix}}):
         chunk_size=None,
         strategy=None,
         intp_t k=1,
-        bint train=False,
     ):
         super().__init__(
             datasets_pair=datasets_pair,
@@ -75,12 +72,18 @@ cdef class ArgKminLRD{{name_suffix}}(ArgKmin{{name_suffix}}):
         )
 
         self.lrd = np.zeros(self.n_samples_X, dtype=np.float64)
-        self.factor = 1 if train else 0
+
+        # Current implementation of ArgKminLRD is used in the fit method
+        # of the LocalOutlierFactor. When it is called from fit the
+        # number of neighbors are actually set to one more than the actual
+        # number of neighbors. There we would actually get k + 1. However
+        # the first neighbor is actually ignored when doing the computation.
+        # This is why the number of neighbors below is set to k - 1.
         self.actual_distances = np.zeros(
-            (self.n_samples_X, self.k - self.factor), dtype=np.float64
+            (self.n_samples_X, self.k - 1), dtype=np.float64
         )
         self.actual_indices = np.zeros(
-            (self.n_samples_X, self.k - self.factor), dtype=np.intp
+            (self.n_samples_X, self.k - 1), dtype=np.intp
         )
 
     cdef inline void _compute_local_reachability_density(
@@ -99,7 +102,10 @@ cdef class ArgKminLRD{{name_suffix}}(ArgKmin{{name_suffix}}):
             num_threads=self.effective_n_threads,
         ):
             overall_reachability = 0
-            for neighbor_rank in range(self.factor, self.k):
+            # Since the first neighbor is ignored when calling ArgKminLRD from
+            # the fit method of the LocalOutlierFactor we simply start the loop
+            # from 1.
+            for neighbor_rank in range(1, self.k):
                 neighbor_idx = self.argkmin_indices[sample_index, neighbor_rank]
                 dist_sample = self.datasets_pair.distance_metric._rdist_to_dist(
                     fmax(self.argkmin_distances[sample_index, neighbor_rank], 0.0)
@@ -109,10 +115,15 @@ cdef class ArgKminLRD{{name_suffix}}(ArgKmin{{name_suffix}}):
                 )
                 dist_reach = fmax(dist_sample, dist_neighbor)
                 overall_reachability += dist_reach
-                self.actual_indices[sample_index, neighbor_rank - self.factor] = neighbor_idx
+                # For actual indices and actual distances we need to fill the
+                # arrays starting from neighbor index = 0 thus we subtract 1
+                # here.
+                self.actual_indices[sample_index, neighbor_rank - 1] = neighbor_idx
                 if return_distance:
-                    self.actual_distances[sample_index, neighbor_rank - self.factor] = dist_sample
+                    self.actual_distances[sample_index, neighbor_rank - 1] = dist_sample
 
+            # self.k - 1 is the actual number of neighbors ignoring the
+            # first extra one.
             overall_reachability = overall_reachability / (self.k - 1)
             self.lrd[sample_index] = 1.0 / (overall_reachability + EPS)