… sklearn/metrics/tests/test_pairwise_distances_reduction.py