@OmarManzoor 2023-05-15
> @OmarManzoor: you can use [asv](https://asv.readthedocs.io) to easily perform benchmarks between those revisions.
> 
> You can take some inspiration from [those ASV benchmarks definitions](https://github.com/scikit-learn/scikit-learn/blob/2c842bd9fac745cdbf98be6ac868553ace06d7ac/asv_benchmarks/benchmarks/pairwise_argkmin_estimator.py) and adapt them for `LocalOutlierFactor`.

Right thank you.