@OmarManzoor 2023-05-16
> @jjerphan Could you kindly have a look if this benchmark seems correct [argkmin_lrd_benchmark](https://gist.github.com/OmarManzoor/87522a2d8b1cbc59dd8620d45cf22851)?

I think this is going to take a lot of time to run.