@OmarManzoor 2023-05-16
@jjerphan @Micky774 These are the results that I got so far

For scikit-learn commit 881123aa <pairwise_distance_backend_for_lof> (round 1/1):
argkmin_lrd.ArgKminLRDBenchmark.time_fit

```
========== ============ ============= ==============
--                    n_test / n_features           
---------- -----------------------------------------
n_train    1000 / 100   10000 / 100   100000 / 100 
========== ============ ============= ==============
 1000      19.3±2ms     18.2±0.1ms    18.0±0.1ms  
10000     1.06±0.02s    1.02±0.02s    1.04±0.02s  
10000000     failed        failed        failed    
========== ============ ============= ==============

For scikit-learn commit c5f10c8b <main> (round 1/1):
[argkmin_lrd.ArgKminLRDBenchmark.time_fit
========== ============ ============= ==============
--                    n_test / n_features           
---------- -----------------------------------------
n_train    1000 / 100   10000 / 100   100000 / 100 
========== ============ ============= ==============
  1000     4.48±0.2ms    4.81±0.1ms    4.65±0.1ms  
 10000      142±6ms       146±2ms       147±5ms    
10000000     failed        failed        failed    
========== ============ ============= ==============
```