@Micky774 2023-05-16
> I don't think this looks good. The performance seems to have decreased.

Can you try profiling (e.g. with Linux `perf`) to see what the hotspots are? I'll also take a look soon to see if I can duplicate your results. We'll get this sorted :)