@OmarManzoor 2023-05-17
@Micky774 I tried using cProfile and there are the results for the argkmin_lrd.pyx file
using this simple script
<Details>

<Summary>Code</Summary>

```python
import cProfile
import numpy as np

from sklearn.neighbors import LocalOutlierFactor

n_train = 1000
n_features = 100

rng = np.random.RandomState(0)
X_train = rng.rand(n_train, n_features).astype(np.float32)
cProfile.run("LocalOutlierFactor(n_neighbors=10).fit(X=X_train)")
```

</Details>

```python
ncalls  tottime  percall  cumtime  percall   filename:lineno(function)
1        0.000    0.000    0.000    0.000     _argkmin_lrd.pyx:134(_finalize_results)
1        0.031    0.031    0.074    0.074     _argkmin_lrd.pyx:34(compute)
1        0.000    0.000    0.015    0.015     _argkmin_lrd.pyx:64(__init__)
```