@OmarManzoor 2023-05-17
> Python and native code can be profiled with [`py-spy`](https://github.com/benfred/py-spy) and results exported for [SpeedScope](https://speedscope.app/) using:
> 
> ```
> py-spy record --native -o py-spy.profile -f speedscope -- python ./script.py
> ```
> 
> I would recommend setting thread affinity to only use one. On GNU/Linux, this can be done using [`taskset(1)`](https://man7.org/linux/man-pages/man1/taskset.1.html):
> 
> ```
> taskset -c 0 py-spy record --native -o py-spy.profile -f speedscope -- python ./script.py
> ```
> 
> Profiles' files can be uploaded here so that they be inspected by readers as well.

I think I realized what the issue is with the significant drop. The metric was being set as euclidean which has a specialized implementation when we use kneighbors.