@OmarManzoor 2023-05-17
These are the latest benchmarks

<Details>

<Summary>Code</Summary>

```python
import numpy as np

from .common import Benchmark

from sklearn.neighbors import LocalOutlierFactor


class ArgKminLRDBenchmark(Benchmark):
    param_names = ["n_train", "n_features"]
    params = [
        [1000, 10000, 100000],
        [100],
    ]

    def setup(self, n_train, n_features):
        rng = np.random.RandomState(0)
        self.X_train = rng.rand(n_train, n_features).astype(np.float32)
        self.y_train = rng.randint(low=-1, high=1, size=(n_train,))

    def time_fit(self, n_train, n_features):
        est = LocalOutlierFactor(n_neighbors=10, metric="manhattan").fit(X=self.X_train)
        self.estimator_ = est
```

</Details>

```python

PR - time_fit 

========= ============
 --         n_features 
--------- ------------
n_train      100     
========= ============
  1000    16.9±0.9ms 
 10000     926±20ms  
 100000   1.54±0.2m  
========= ============


MAIN - time_fit

========= ============
--         n_features 
--------- ------------
 n_train      100     
========= ============
   1000    16.5±0.7ms 
  10000     926±9ms   
  100000   1.49±0.01m 
========= ============


```