@OmarManzoor 2023-05-17
Benchmarks with float64

<Details>

<Summary>Code</Summary>

```python
import numpy as np

from .common import Benchmark

from sklearn.neighbors import LocalOutlierFactor


class ArgKminLRDBenchmark(Benchmark):
    param_names = ["n_train", "n_features"]
    params = [
        [1000, 10000, 100000],
        [100],
    ]

    def setup(self, n_train, n_features):
        rng = np.random.RandomState(0)
        self.X_train = rng.rand(n_train, n_features).astype(np.float64)
        self.y_train = rng.randint(low=-1, high=1, size=(n_train,))

    def time_fit(self, n_train, n_features):
        LocalOutlierFactor(n_neighbors=10, metric="manhattan").fit(X=self.X_train)
```

</Details>

```python

PR - time_fit 

========= ============
 --         n_features 
--------- ------------
n_train      100     
========= ============
  1000     18.6±0.1ms 
 10000     1.11±0.02s  
 100000    1.77±0m  
========= ============


MAIN - time_fit

========= ============
--         n_features 
--------- ------------
 n_train      100     
========= ============
   1000    18.7±0.4ms 
  10000    1.11±0.02s   
  100000   1.80±0.01m 
========= ============


```